Bug fixes :bug: and removal of car from garage :car:
Title. There was a bug where the car being passed through to the auction
was coming from the database rather than the user's garage, causing it
to sell the wrong car and therefore not be removed, leading to an
infinite loop of being able to sell a car you don't own. Talk about
illegal!