Successful refactor of Auction.js
The bug with all of the AI bidding at the same time has not been fixed,
but there was a bunch of code being called (from a previous version)
that wasn't being used - it was literally doing nothing. Tested after
the removal and all functionality remains. I was hopeful the refactor
would solve the issue, but alas it has not so the show goes on.