SaleView Headache
I spent my time today screwing around with the Saleview and do you
remember that issue we had with the for loops scope not matching? It
still exists and that problem is present in other parts of the file as
well. I wasn't able to get much going because of it.

I'm going to keep screwing around with it but so far that file is fubar
and I have no clue hoe it still works when notepad is showing the scopes
so fucked up.