Trying to fix ai leaving at same time instances
I believe the issue is with the bid caps being too close to each other
and not to do with the sorting at all. My thought was to check between
ai's bid caps and if their within a certain range adjust them. Ive
failed at that so far, I cant decide if Im wasting my time or not.